+++ tests/fixers/test_utils_timezone.py
@@ -36,13 +36,66 @@ def test_unrecognized_import_format():
     )
 
 
+def test_untouched_utc_attribute():
+    check_noop(
+        """\
+        timezone.utc
+        """,
+        settings,
+    )
+
+
+def test_not_imported_utc_name():
+    check_noop(
+        """\
+        utc = 1
+        utc
+        """,
+        settings,
+    )
+
+
 def test_fixed():
     check_transformed(
         """\
         from django.utils.timezone import utc
+        foo(utc)
         """,
         """\
-        from datetime.timezone import utc
+        from datetime import timezone
+        foo(timezone.utc)
+        """,
+        settings,
+    )
+
+
+def test_fix_skips_other_utc_names():
+    check_transformed(
+        """\
+        from django.utils.timezone import utc
+        utc
+        myobj.utc
+        """,
+        """\
+        from datetime import timezone
+        timezone.utc
+        myobj.utc
+        """,
+        settings,
+    )
+
+
+def test_fix_inner_import():
+    check_transformed(
+        """\
+        def do_something():
+            from django.utils.timezone import utc
+            something(utc)
+        """,
+        """\
+        def do_something():
+            from datetime import timezone
+            something(timezone.utc)
         """,
         settings,
     )